v0.8.13b-a
Hey Alpha Gang (lol)

So little update on what and why that alpha:

To bring out more and consistent updates and to get the feedback from all of you while the game is in development, instead of individual BIG updates in the beta with a "swarm" of bugs that need to be fixed because of all the new updates. (this way will also be less stressful coding for me)

So this is why the alpha version is reintroduced! For more regular updates, but also less polished.

The code you guys need to unlock the fan tank on your account (https://weetanks.com/redeem)

[hah you thought]

You can now switch to v0.8.13b-a on Steam.

Changelog:
- Fan Tank added
- Custom Tank added RD missiles (in tab 2)
- Custom Tank menu live preview of the tank, instead of the image, you can spin the tank!
- You can now set difficulty for testing before starting
- Add bomber plane to survival mode, that starts appearing after wave 40
- Make tank dying sound moddable
- Sign in status on main menu fixed so that its actual
- Skins from in-game work again (they now override modded skins)
- Electro pads super glow fixed
- Enemy wall bounce shots are changed, instead of scanning like a lighthouse it now scans randomly, for more unique shots
- Tanks can now shoot in between looking angles, this might have big impact on gameplay!! Needs testing. (this also prevents the burst mode from custom tanks)
- Fixed shotgun custom tanks not counting bullets right
- Fixed So Close achievement not being achievable
- Bomber plane should stop spawning when king tank is ded,* I havent tested out the new code yet tho.*
- Garbage quality is now even more garbage
- Custom tank limit is 40
- Mission limit is now 100 (mission 100 might be bugged because of some king tank scripts)
- Map editor auto save added!
- Players can now shoot silver bullets in the map editor
- When signing in , it should not ask for transfer steam account anymore if the account is already linked to a steam account
- You can now overwrite an old account that is linked to your steam ID with a newly made account
- There is a "set new password" button in the account menu for automated steam accounts